Move CachingEntityRepo to own file
